input: cypress touchkey: remove unused firmware loading code
The firmware loading code is not used by any of the devices
currently supported by Replicant.

Some of the files (issp_*) that are involved in the firmware
loading have unclear licensing statements.

Removing the problematic files and adapting the code to compile
without them is easier and faster than getting legal advise to
understand if the files are under a free software license or not.

Change-Id: I611a4b900cdea43b7fcf7823a1c9a438593c59e2
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
[ TALU - Applying simply for cleanup purposes ]
Signed-off-by: TALU <taluemail272@gmail.com>