Implement some additional CPU governors.
These governors are from the DorimanX kernel here: https://github.com/gsstudios/Dorimanx-SG2-I9100-Kernel The governors had to be modified to run on this kernel, and everything relating to earlysuspend had to be removed/commented. Governors added are intellidemand, HYPER, darkness, nightmare, zzmoove and neox. Sleepy couldn't be implemented as it relies on earlysuspend. All of the pre-existing governors will work the exact same as before.

Change-Id: Iea35536b9faafd7b4dbaa31b339256efed17aca4