max8997-muic: properly fix build warning in max8997_muic_mhl_detect.
Charging detection didn't work properly after a build warning "fix" because in the max8997_muic_mhl_detect function, there was a mutex_lock for &info->mutex, and later on in the 'out' label, there was a mutex_unlock. The previous warning fix meant that the 'out' label was only included if CONFIG_MACH_U1 and CONFIG_MACH_TRATS were not defined, so there was never a mutex_unlock. I'm not sure how the contents of the 'out' label were actually used even when CONFIG_MACH_U1 and CONFIG_MACH_TRATS were not defined as the only place the 'out' label was called to was in code compiled only when both are not defined. The fix is to do a mutex_lock only if both are not defined in the first place.

Change-Id: I76e9a0d793d1494f0a51e9d81131cc0cdcb4531c